failing c003

+++ src/challenge/c001.rs
@@ -1,7 +1,7 @@
 use crate::util::is_divisible;
 
 #[allow(dead_code)]
-pub fn solve(a: i32) -> i32 {
+pub fn solve(a: i64) -> i64 {
     let mut m = 0;
     for n in 1..a {
         if is_divisible(n, 3) || is_divisible(n, 5) {
