+++ src/util/mod.rs
@@ -1,4 +1,4 @@
-pub fn is_divisible(n: i32, k: i32) -> bool {
+pub fn is_divisible(n: i64, k: i64) -> bool {
     n % k == 0
 }
